+++ src/vm/wren_value.c
@@ -241,8 +241,9 @@ ObjForeign* wrenNewForeign(WrenVM* vm, ObjClass* classObj, size_t size)
   return object;
 }
 
-ObjFn* wrenNewFunction(WrenVM* vm, ObjModule* module, uint8_t maxSlots)
+ObjFn* wrenNewFunction(WrenVM* vm, ObjModule* module, uint16_t maxSlots)
 {
+  
   ObjFn* fn = ALLOCATE(vm, ObjFn);
   initObj(vm, &fn->obj, OBJ_FN, vm->fnClass);
